- Bring asset response type into shared package
