+++ packages/wallet/frontend/src/pages/account/create.tsx
@@ -18,7 +18,7 @@ import { Controller } from 'react-hook-form'
 import { NextPageWithLayout } from '@/lib/types/app'
 import { useOnboardingContext } from '@/lib/context/onboarding'
 import { useEffect } from 'react'
-import { createAccountSchema } from '@wallet/shared/src/responses'
+import { createAccountSchema } from '../../../../shared/src/types'
 
 type CreateAccountProps = InferGetServerSidePropsType<typeof getServerSideProps>
 const CreateAccountPage: NextPageWithLayout<CreateAccountProps> = ({
