+++ packages/wallet/shared/src/types/index.ts
@@ -1,1 +1,2 @@
 export * from './account'
+export * from './asset'
